create shared metadata file
