+++ tests/src/constants.ts
@@ -9,3 +9,8 @@ const constants = {
 } as const;
 
 export default constants;
+
+export enum MetadataAppCommand {
+  sendGetInfoMetadataToTests = "sendGetInfoMetadataToTests",
+  confirmRaisedIntentReceived = "confirmRaisedIntentReceived",
+}
